+++ collection/cluster.go
@@ -21,8 +21,8 @@ import (
 	"strings"
 
 	"github.com/dremio/dremio-diagnostic-collector/cli"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/root/helpers"
 	"github.com/dremio/dremio-diagnostic-collector/cmd/simplelog"
-	"github.com/dremio/dremio-diagnostic-collector/helpers"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/masking"
 )
