+++ collection/collector_test.go
@@ -25,7 +25,7 @@ import (
 	"time"
 
 	"github.com/dremio/dremio-diagnostic-collector/cli"
-	"github.com/dremio/dremio-diagnostic-collector/helpers"
+	"github.com/dremio/dremio-diagnostic-collector/cmd/root/helpers"
 	"github.com/dremio/dremio-diagnostic-collector/pkg/tests"
 )
